+++ net/buffer.c
@@ -86,7 +86,7 @@ size_t buffer_write_avail(buffer_t* buffer)
 size_t buffer_read(buffer_t* buffer, unsigned char* buf, size_t len)
 {
 	size_t to_read = min(len, buffer_read_avail(buffer));
-	for (int i = 0; i < to_read; i++) {
+	for (size_t i = 0; i < to_read; i++) {
 		buf[i] = buffer->data[(buffer->read_ptr + i) % buffer->real_size];
 	}
 	buffer->read_ptr = (buffer->read_ptr + to_read) % buffer->real_size;
@@ -107,7 +107,7 @@ size_t buffer_write(buffer_t* buffer, const unsigned char* buf, size_t len)
 {
 	size_t to_write = min(len, buffer_write_avail(buffer));
 	int write_ptr = (buffer->read_ptr + buffer->read_avail) % buffer->real_size;
-	for (int i = 0; i < to_write; i++) {
+	for (size_t i = 0; i < to_write; i++) {
 		buffer->data[(write_ptr + i) % buffer->real_size] = buf[i];
 	}
 	buffer->read_avail += to_write;
